Have Cider plugin be part of :user, not :repl
The Cider docs say make cider-repl a plugin only for the :repl profile, or else the middleware will run all the time and delay startup.  It says you only need it for :repl.  But that's not really true.  If you want to run the project with "lein run" and connect Cider to it, you need to get the middleware plugged in.  The options are to include it in project.clj, or have it more prominent in profiles.clj.  This seems like a better option.