Assume case-insensitive file system
When doing the first round of moves, we check to see if the file is already in its best position.  Lower-case both filenames for the comparison, so it becomes case-insensitive.

For the second (deferred) round of moves, we were not checking at all if we had hit the best option.  Do the check, but again, convert both names to lower case first.

Update a comment.