Add file_exists method
file_exists() is generally nothing more than os.path.exists().  The primary reason it merits another function is to support the "domove == False" functionality.  In that mode, we are supposed to not move any files, but to output exactly which moves we WOULD do.  But that's not so straightforward, because the moves we would do depend on the earlier moves we would have done.

A secondary reason to have a separate function is for debug logging.