Be explicit about iterating over keys
When we iterate over the list of moves, we are just accessing the keys.  So don't iterate over the dictionary itself, which gives us the values that we don't bind; get the keys, store in a local variable, and iterate over that.

Also, "cache" the length of the dictionary into a local variable before the loop.