Destructively modify moves list
Instead of treating the initial dict as immutable, and producing a separate dict of deferred moves, modify the dict in place.  When we do the move, delete the key from the dict, and then at the end of one iteration, the deferred moves are all that is left in the dict.

In order to do this without Python getting confused, make a copy of the keys.