Use my-shell instead of shell
Add my-shell function, which goes to the most recently used shell buffer, in the existing buffer, moves point to the end of the buffer, and recenters it.

This takes the key binding previously used by 'shell, which goes to the buffer named *shell*, does not move point, does not recenter, and, as of more recent versions of emacs, does not necessarily use the same buffer.